refactor: Upgrade mainactivityviewmodel
