+++ app/src/test/java/uk/gov/onelogin/MainActivityViewModelTest.kt
@@ -6,9 +6,9 @@ import org.junit.jupiter.api.Test
 import org.mockito.kotlin.mock
 import org.mockito.kotlin.verify
 import uk.gov.android.network.api.v2.ApiRequest
-import uk.gov.android.network.api.v2.ApiResponse
-import uk.gov.android.network.api.v2.ApiResponseAssertions.expectFailure
-import uk.gov.android.network.api.v2.ApiResponseAssertions.expectSuccess
+import uk.gov.android.network.api.v3.ApiResponse
+import uk.gov.android.network.api.v3.ApiResponseAssertions.expectFailure
+import uk.gov.android.network.api.v3.ApiResponseAssertions.expectSuccess
 import uk.gov.android.network.attestation.ClientAttestationProvider
 import uk.gov.android.network.attestation.TestClientAttestationProvider
 import uk.gov.android.network.auth.AuthenticationProvider
@@ -18,7 +18,7 @@ import uk.gov.android.network.client.v2.GenericHttpResponse
 import uk.gov.android.network.client.v2.StubHttpClient
 import uk.gov.android.network.dpop.DPoPProvider
 import uk.gov.android.network.dpop.TestDPoPProvider
-import uk.gov.android.network.service.DefaultNetworkService
+import uk.gov.android.network.service.v2.DefaultNetworkService
 
 class MainActivityViewModelTest {
     private val genericHttpClient: GenericHttpClient = mock()
@@ -55,7 +55,7 @@ class MainActivityViewModelTest {
 
         val success = defaultNetworkService.makeRequest().expectSuccess()
         assertEquals(
-            ApiResponse.Success(okResponse.body, okResponse.status),
+            ApiResponse.Success(okResponse.status, okResponse.body),
             success
         )
     }
